Fix and complete documentation for master table
Fix the documentation for the master table implementation. It is noted
that there needs to be some work put into tagging functions as private
in documentation, and which methods should be used by the user are not
presently clear. Better yet, private methods should be removed from
headers, if possible.