Change test_skiplist to use sizeof(int)
Before it used to assume a fixed size of 4 bytes for an integer. My theory here is that this caused the unit tests to copy in garbage bytes on device, which uses 2 byte ints. This was then throwing off the comparison functions, since it was comparing into garbage data (tehnically initalized garbage data, but garbage nonetheless).