Change ion MT delete behaviour
Delete master table (i.e. nuke the roster) now works even if the table is not open. This makes sense, since we don't want to bother having to open the master table only to immediately delete it (in the case where you want to just restart fresh, regardless of what the previous state is.)